+++ game_api/gameLIB/components/stage/instructions/Instruction.cs
@@ -8,9 +8,9 @@ namespace gameLIB.components.stage.instructions
     public abstract class Instruction
     {
         public String type { get; set; }
-        public int id { get; set; }
+        public ulong id { get; set; }
 
-        public Instruction(String _type, int _id)
+        public Instruction(String _type, ulong _id)
         {
             type = _type;
             id = _id;
