+++ game_api/gameLIB/components/stage/instructions/InstructionShoot.cs
@@ -9,9 +9,9 @@ namespace gameLIB.components.stage.instructions
         public Vector2 direction { get; set; }
         public String particleType { get; set; }
         public Vector2 destination { get; set; }
-        public int particleId { get; set; }
+        public ulong particleId { get; set; }
 
-        public InstructionShoot(int _id, String _particleType, int _particleId, Vector2 _destination, Vector2 _direction, float _speed)
+        public InstructionShoot(ulong _id, String _particleType, ulong _particleId, Vector2 _destination, Vector2 _direction, float _speed)
             : base(String.Empty, _id)
         {
             direction = _direction;
